Refactoring board routes
